refactor: split admin GraphQL into stitched subgraphs
