+++ package.json
@@ -20,6 +20,9 @@
     "rebuild:indexer": "tsx scripts/rebuild-indexer.ts"
   },
   "dependencies": {
+    "@graphql-tools/schema": "^10.0.38",
+    "@graphql-tools/stitch": "^10.2.2",
+    "@graphql-tools/utils": "^11.2.2",
     "@openfeature/server-sdk": "^1.22.0",
     "@opentelemetry/api": "^1.9.1",
     "@opentelemetry/exporter-trace-otlp-http": "^0.57.2",
@@ -70,5 +73,6 @@
     "typescript": "~5.6.2",
     "vite": "^8.1.0",
     "vitest": "^4.1.7"
-  }
+  },
+  "overrides": {}
 }
