+++ test-schema.ts
@@ -1,0 +1,2 @@
+import { gatewaySchema } from './src/graphql/gateway.js';
+console.log('Schema built successfully');
